Links display when cards are clicked but going back after clicking a link causes an error
