+++ src/components/PersonalPage/PersonalPage.jsx
@@ -4,14 +4,15 @@ import "./PersonalPage.css";
 import {} from "module";
 import profileIcon from "../../assets/profile-icon.png";
 import UserLinks from "./Links/UserLinks";
+import { useEffect } from "react";
 
 
 const PersonalPage = ({ users }) => {
   let { userId } = useParams();
-
-  // const user = users.find(user => user.name === userId);
-  // console.log(user)
-  // const links = user.links;
+  const user = users.find(user => user.name === userId);
+ 
+  const links = user.links;
+  console.log(links)
 
   // const handleClick = (url) => {
   //   window.open(url);
@@ -31,7 +32,7 @@ const PersonalPage = ({ users }) => {
       </div>
       {/* <div>{buttons}</div> */}
       <div>
-        {/* <UserLinks links={links} /> */}
+        <UserLinks links={links} />
       </div>
     </main>
   );
@@ -39,18 +40,10 @@ const PersonalPage = ({ users }) => {
 
 
 PersonalPage.propTypes = {
-  users: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.string.isRequired,  // Ensure id is included and required
-      name: PropTypes.string.isRequired,
-      image: PropTypes.string,
-      type: PropTypes.string,
-      occupation: PropTypes.string,
-      description: PropTypes.string,
-      skills: PropTypes.arrayOf(PropTypes.string),
-      links: PropTypes.arrayOf(PropTypes.string).isRequired  // Ensure links is included and required
-    })
-  ).isRequired,
+  user: PropTypes.shape({
+    name: PropTypes.string.isRequired,
+    links: PropTypes.arrayOf(PropTypes.object).isRequired
+  }),
 };
